modifique en el modelo en la clase receta la columna de cantidad para que sea booleano.

+++ EmpleadosEBS/Migrations/20190820192939_Inicial.cs
@@ -234,7 +234,7 @@ namespace EmpleadosEBS.Migrations
                         .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
                     ArticuloID = table.Column<int>(nullable: false),
                     PlatoID = table.Column<int>(nullable: false),
-                    Cantidad = table.Column<int>(nullable: false)
+                    Cantidad = table.Column<double>(nullable: false)
                 },
                 constraints: table =>
                 {
